Fix how settings link is added on plugin page
Previously the "Settings" link was not showing up under the plugin because we were using
the old way of adding links to the plugin page. Breaking things out into different files broke the
old way of checking for adding the "Settings" link since the main plugin file was no longer the same
as the file which produced the link. Now we're registering the callback under the name of the plugin
(rather than hooking into `plugin_action_links` in general) so we don't need to gaurd based on file
name.